Initial commit
Work with the Flask library to implement a web application that will allow you to keep an eye on the weather. You will use Flask-SQLAlchemy to connect a database to your web application and work with a weather service API to get the current weather from all over the world. Practice HTML templates with Jinja and handling requests.